+++ client/package.json
@@ -50,5 +50,7 @@
     "nodemon": "^1.18.10",
     "prettier": "^1.16.4"
   },
-  "peerDependencies": {}
+  "peerDependencies": {
+    "web3": "^1.0.0-beta.52"
+  }
 }
